libvirt: display warning message if firewalld is enabled
Libvirt versions < 1.2.4 are very inefficient at talking to
firewalld. This results in slow VM startup/shutdown which
in turn causes tempest to often time out.

Display a warning if firewalld is enabled on distros which
commonly have this inefficient combination enabled, so that
developers can decide whether to disable firewalld or to
update their libvirt or ignore the problem.

Change-Id: I1dc610dfb8966e6e4e70c19416ee85e332bde9f2