Update patch set 5
Patch Set 5:

f20+default libvirt+firewalld is *not* a valid solution.  If it can't run tempest in under 2 hours by default -- out-of-the-box without touching anything -- it is just as broken as if it were segfaulting.   devstack can not default to a case where tempest takes more than two hours because you happen to never run it.  nobody reads warnings

nobody but me has seemed to notice this so it's a storm in a tea-cup anyway.  Or perhaps people tried fedora, their CI took 4 times as long as Ubuntu so they just said "fedora is crap" and moved on (who'd blame them)

Patch-set: 5