Update patch set 1
Patch Set 1:

Fundamentally being able to test & develop openstack with firewalld enabled *is* a valid & necessary development scenario to support. Unconditionally blocking use of devstack is not an acceptable solution, because it is taking the decision out of the hands of the developer. devstack is here to serve the needs of the developers.

Those developers who care about tempest are more than intelligent enough to see the warning message & deal with it if they so desire.

Patch-set: 1