Update patch set 7
Patch Set 7:

'something like this' should be interpreted as 'I didn't thoroughly test this'  ;)

Inside a VM with 1 vCPU this sets the worker count to 0, does celio do the right thing then?  Really, we should check for that and not subject the app to such rudeness.

 $(( ($(grep ^processor /proc/cpuinfo | wc -l) + 1) / 2 ))

will ensure that we don't try to choke off all workers.

Patch-set: 7
Label: Verified=0