Use 8GiB swift storage with tempest
"No space left on" appeared in the swift proxy log,
increasing the size to 8GiB.

Several tempest test using 2 GiB swift storage.
The original calculation made with 1 GiB.

The tempest test will be changed to reduce the theoretical
peak usage, but 8GiB will be required when we increase the
number of workers.

This simple change expected to solve or mitigate
 the leading gate failure.

Related-Bug: #1347692

Change-Id: I11c060fe5d462fa9ef37348be75d6125868529be