Fix MySQL password on Ubuntu
Ubuntu relies on debconf to set the MySQL server password when the MySQL
package is installed. (This is done to avoid the password prompt during
install).

This works the first time `./stack.sh` is run.

On subsequent runs, however, the package will already be installed, so the new
password won't be set.

The proposed work around is to uninstall/reinstall mysql-server on each run.

Change-Id: I046ae22d59a99ef86da786c4d5db870814ce159f
Closes-Bug: 1358485