Update patch set 1
Patch Set 1:

> The password should have been generated and saved in localrc. So that was removed?

No it's saved correctly to localrc. No problem there.

The problem is that the password is only set for MySQL on the first-run of stack.sh, when it installs the MySQL package.

Subsequent runs that don't install the MySQL package will generate the password, but be unable to actually set it.

> i don't agree with purging it every run though

Agree that this approach isn't ideal, but it does work at least.

Any alternative proposals?

Patch-set: 1
Reviewer: Gerrit User 475 <475@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0