Update patch set 1
Patch Set 1:

>Sorry, I think we're mis-understanding each other.

Yeah, few things going on here, so think we're focusing in on different things :-)

> The password will be generated and stored in localrc (stack.sh:read_password)
For the next run, that same value should be used. 

If you have a `localrc`, it'll use the same password each time, but if you don't and instead use `.localrc.auto` , it'll be different between runs.

>The password should not be regenerated if localrc has not been removed.

If the `localrc` is removed, then it's going to prompt for new passwords for all of the services: why shouldn't MySQL take on a new password, like the other services do?

Patch-set: 1