Update patch set 1
Patch Set 1:

> if you ...use `.localrc.auto` , it'll be different between runs.

> If the `localrc` is removed, then it's going to prompt for new passwords for all of the services: why shouldn't MySQL take on a new password, like the other services do?

unstack.sh has in it

  # mysql and rabbit are left running as OpenStack code refreshes
  # do not require them to be restarted.

so there's really a precedent to keep them around -- i.e. not purge the database on people.  

Just below this bit, the password is being saved into .my.cnf?  i think it would be reasonable to assume that if the db password is not set, and .my.cnf exists, then the password should come from that before generating a new one?

Patch-set: 1