Update patch set 1
Patch Set 1:

> so there's really a precedent to keep them around -- i.e. not purge the database on people.

Yeah, that's a fair point.

Still, without a fix, we're left in a situation where devstack can prompt you for a new password and then not actually perform the set-password in MySQL. Even worse, the code to GRANT privileges expects the new-password to be set.

So I think we should do /something/ to fix this, just not sure what (if not this)...

[Will think about it some more tonight]

Patch-set: 1