Update patch set 7
Patch Set 7:

danpb, the response was (and is): DevStack is not in the business of configuring networking more than what is required for Nova-Net or Neutron.  We'll leave well enough alone except for addressing brokenness (this review) as needed.

Can you show me how to undo the network config effects of launching a server using either Nova-Network or Neutron?

Patch-set: 7