Update patch set 7
Patch Set 7:

When Docker was still supported we required them to do the basic Docker installation and configuration before stack.sh ran.  It was in a script in tools so it was in the repo, but otherwise DevStack stayed out of preparing the system.  OpenDaylight was in a similar situation, Kyle and I came up with a split that seems to have worked.  On the other hand, Ceph is completely installed from within lib/ceph when it is enabled.

As you can see, there isn't a lot of consistency in regard to extra-curricular services.  But there was a demand for all of them in one way or another.  I do not see any demand (before your request) for adding basic network firewall configuration to DevStack, nor do I see why it should be included.  If Nova-Net or Neutron needed it that would be different.  And if there was, it would be in an additional review anyway.

Patch-set: 7