Change API_RATE_LIMIT default to False to match the nova code
The api_rate_limit config option in nova has defaulted to False since
Havana per commit cfdc59860ef1482e3b064f90f719ec9e484590ce.  This change
simply updates the default value in devstack to match the default in
Nova.

Note that this variable exists in devstack essentially for Tempest so we
can set it in devstack-gate, which is kind of a moot point now given the
code defaults to it anyway.  I'm leaving it here in case anyone is using
devstack and relying on setting api_rate_limit=True via the variable.

Change-Id: Ib0538bdd23b17e519b9c917018ccc9fa8c6425c5