Update patch set 1
Patch Set 1:

This is one of a number of knobs that pre-date the capabilities to set arbitrary values in conf files via local.conf.  I would prefer to remove it entirely as it is not used anywhere except to set the value in nova.conf.  Anyone depending on the default being true is broken by this change anyway, the only difference is the config change they would need to make.

Patch-set: 1
Label: Verified=0