Update patch set 1
Patch Set 1:

Ian, I haven't tried on RHEL 7, but if you're going to be doing development on nova, keystone or any other project that gates on the docs, you'll need to run sphinx which means you'll be pulling in graphviz.

I'm not sure if there is a different more slimmed-down package that installs the dot command that sphinx is trying to use on RHEL 7, if there were we could use that instead.

Patch-set: 1
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0