Update patch set 3
Patch Set 3:

Why are we going to introduce additional gate jobs?

My preference would be to just change what we test with. This keeps the current test matrix to the existing size and will allow us to avoid needing to track bugs for multiple drivers.

I think we keep trying to test all the things and it is only additional burden on the folks that eventually debug the many sporadic failures that we see.

Patch-set: 3