Update patch set 1
Patch Set 1: Code-Review-2

We can't commit this until the horizon support is done in [1], right?  i feel that's a -2 blocker until that is merged, as the option is broken until then

but I fairly strongly feel we should enable some non-md5 default for this -- otherwise, it seems very ripe for bit-rot.  we add services and forget this, everything defaults to md5 and we never notice.

[1] https://review.openstack.org/#/c/116509/

Patch-set: 1
Label: Code-Review=-2