Update patch set 2
Patch Set 2: Code-Review-1

Not sure if the jenkins failures are related, didn't look into that

A priori this seems like a weird thing to do.  I presume this is multi-node where you've got these services on another host, and you're not running tempest there?  can you please just expand in the description what scenario this is useful for

Patch-set: 2
Label: Code-Review=-1