Update patch set 2
Patch Set 2:

Hi Sean, thanks for the feedback.

I'll certainly pull the PPA.  While the required packages are not yet available via the proper ubuntu channels, they are available via EPEL, which - if I understand correctly - is enabled in devstack by default.  I'd be happy to rework the patch to remove the PPA and only support AMQP1.0 on epel-compatible distros until ubuntu support arrives.

However, the purpose of this change is greater than just satisfying the proposed gate check. I probably haven't done a good job of explaining the 'big picture'.  AMQP 1.0 is the primary focus of the Apache QPID team going forward.  This change, along with the new AMQP 1.0 oslo.messaging driver, is part of an effort to encourage QPID deployments towards this new protocol standard, and away from the deprecated 0-10 protocol.

So aside from the gate test, this change has value for those deployments that want to adopt the 1.0 protocol that is supported by the latest QPID broker.

Patch-set: 2
Label: Verified=0