Update patch set 2
Patch Set 2:

Ken - there is no schism over support that would cause divergent implementations, right?  

Assuming everyone is working to the same goal, my feeling is that with full support on a range of platforms (basically all rpm based distros), then we shouldn't hold things up.  We can be doing real gate-testing of this feature on a f20 node

When ubuntu support is available, it should be in much better shape for the intervening testing that has been done

Patch-set: 2