Update patch set 1
Patch Set 1: Code-Review+1

If projects are gating on docs, then it seems to me that needs to be testable and dependencies need to be in the requirements.  While I agree with Ian that it would be nice to not drag in X, I don't see a way around this.

Patch-set: 1
Reviewer: Gerrit User 7175 <7175@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1