Update patch set 2
Patch Set 2: Workflow-1

Sorry, folks - I forgot to take my Anti-Stupid pills this a.m.

I'm marking this patch a WIP for now.  I've hit a problem with this patch on Centos6.  While the AMQP 1.0 libraries are available on EPEL for Centos6, the version of qpidd is -way- old, and doesn't have AMQP 1.0 support :P

Centos7 and Fedora 20+ won't have this issue, as both use EPEL to get the latest libraries and qpidd.

Patch-set: 2
Label: Workflow=-1