Update patch set 5
Patch Set 5:

@Ian, sorry about that. I'll work on documenting it better.

Glance store is a new library the Image Service program is working on. It contains the code that used to live under `glance/store` which is responsible for managing the stores where Glance stores the data into.

In order to migrate Glance to this library, we need to remove the `glance/store` package from the code base but in order to do that, we need this patch to land as well.

Does the above clears your doubts?

The Glance team has agreed on moving forward with the library and all the necessary patches have been submitted. I believe/hope, this is the last one required outside Glance.

Patch-set: 5