Update patch set 2
Patch Set 2:

I see that your gate test on fedora 20 [1] is using the functional tests under tox, i.e. this devstack change is not for use in the gate.

I think installing the ppa is helpful in getting the feature enabled, as long as Ken is willing to keep an eye on it and clean-up when appropriate.

My +2 is if [1] gets merged as a tox job, then I think this is OK for devstack 

[1] https://review.openstack.org/#/c/115752/

Patch-set: 2