Update patch set 5
Patch Set 5:

I don't think the solution is to rename glance.store to glancestore. That's confusing and it's not the way things should work. It's also really late in the game to change the name of the package and lots of things have been done to make this happen in Juno.

Given the fact that the project has been created in openstack, all the other patches have been merged and we've also agreed on going with namespaces. Can we let this one land? (which is hopefully the last patch outside Glance we need)

Patch-set: 5