Update patch set 4
Patch Set 4: Code-Review+1

I dislike that is_ubuntu is true for a debian system, but debian is not an ubuntu system; ubuntu is a debian system.

That said, changing these functions is beyond the scope of this patch.  That being the case, the methodology of this is acceptable.

Patch-set: 4
Label: Code-Review=+1