Update patch set 5
Patch Set 5:

There is more to 'supported' as Ian mentions, regular testing is necessary but insufficient.  As the docs say, Ubuntu LTS and latest release, Fedora latest and N-1 are our supported distros.  Others are allowed within those two families (hence the names of is_ubuntu and is_fedora) but I do not consider them supported in the same sense.  Consider the state of Suse support.  Debian and RHEL are included because they are in the same family, ignoring the parent/child relationships.

Patch-set: 5
Label: Verified=0