Reset no_proxy when the node is installed behind a proxy
Closes-Bug: 1291111/1322200

Problem:
=======
When devstack is installed behind a proxy with proxy variables
like http_proxy and https_proxy exported and without no_proxy
set, the installation will fail. The reason is that any REST
request sent to Keystone will be redirected to the proxy server.

Solution:
========
Fix involves appending no_proxy variable with localhost,HOST_IP,
FLOATING_RANGE and FIXED_RANGE if they are not included and
exporting the variable.

Change-Id: If65cb282adc6f5e4c4507b1422885d81de3da4e8