Skip SEGMENTATION_ID checking with flat provider network
Currently, devstack can't deploy OpenStack with flat provider network.

devstack checks whether SEGMENTATION_ID is provided or not when it
deploys openstack with provider network. If it is not provided devstack
prints error message and dies.
On the other hand, devstack also gets error from "neutron net-create"
command even if it sets SEGMENTATION_ID with flat option.

To fix this issue, this patch just skip checking SEGMENTATION_ID if
PROVIDER_NETWORK_TYPE is set to "flat".

Change-Id: I2273e3369c688bbce1a6e39b43635b5548cc2b98
Closes-Bug: #1362382