Update patch set 1
Patch Set 1: Code-Review-1

I'm negative on this but willing to be out-voted, because I think it's a symptom of wrong behaviour.

IMO the fact worlddump gets dumpped into the root directory is wrong ... if $LOGDIR is set it should go into the logs directory (that currently happens) and if it isn't, it should just come out on stdout and not write a log file.

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1