Update patch set 9
Patch Set 9: Code-Review-1

(2 comments)

Please take a look on comments. It seems that you expect that all NoSQL dbs has 'database' as SQL does, kind of 'namespace'. But it's not the case with HBase. But maybe my suspicion is wrong :)
Anyway, the main question is about recreate_nosql_database method. Why do we need that? E.g. in Ceilometer we need several tables with different cfs. We cannot know that on devstack lvl

Patch-set: 9
Label: Code-Review=-1