Update patch set 1
Patch Set 1:

Sean,

There is a little bit of sensibility to the paths even if they don't make sense at first:

imagebuild/ contains directories that can build IPA images. The only one currently implemented is CoreOS, in imagebuild/coreos. Building the CoreOS image outputs the image to imagebuild/coreos/UPLOAD/ (which was initially chosen to indicate that contained the image to upload).

If this were to change, it would break our check because it would break devstack.

Patch-set: 1
Label: Verified=0