Update patch set 15
Patch Set 15: Code-Review-1

(16 comments)

Functionally this works fine for me, so the logic and operations are good.

I would like to see improvements in the comments and layout to make the code easier to read. I think many people who want to experiment with IPv6 will be looking at this as a guide.

Patch-set: 15
Label: Code-Review=-1