Update patch set 6
Patch Set 6: Code-Review-1

I'm a bit negative on this, but willing to be out-voted.

It seems like we are already missing things; heat, ceilometer, neutron all have worker options meaning various things.  This is yet-another-flag with fuzzy semantics and name-space; e.g. neutron has an api_workers option that doesn't fit with this, but yet we've got a global "API_WORKERS" variable introduced with this change.  It seems these settings are the type of things that should be globally left to upstream defaults and tweaked locally in localrc as required for various environments

Additional documentation on tips for running devstack in low-memory/high-cpu environments would be a welcome alternative to this IMO

Patch-set: 6
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1