Update patch set 6
Patch Set 6:

I'm very keen for this to be a devstack setting in localrc; my use case is for a 3rd party CI where devstack-gate only supports localrc and so there is currently no way of setting any maximum number of workers.  In an environment with 6 CPUs (so not even a 'large' number) we have at least 12 workers more than are needed (Tempest runs with a maximum of CPU/2 threads).  Each worker seems to take 3% of 4G (120M) and reducing this number has a notable impact on the memory usage in the CI environment. 

I'd be very open to a name change to restrict the scope, or to expand it to include other workers that you think should be included in this.  Suggestions very much welcome.

Patch-set: 6
Label: Verified=0