Move apache config creation to start_keystone()
We currently cleanup keystone apache config during stop_keystone
when mod wsgi is enabled, but only create it during initial keystone
configuration.  Grenade expects to be able to stop and start the
service without reconfiguring it, using only the respective functions.
This moves rendering of the keystone apache site config to the start
function.

Change-Id: I9e57c9a763fbdb8c5737decfe797ba0e9894150b
Partial-bug: #136515