Update patch set 6
Patch Set 6:

I don't dispute the problem, but I still feel this flag is too abstract about what it is setting (and not setting) and thus is not a good choice for a global.

the devstack-gate scripts have extant pre-test hooks where you could setup local.conf values before the stack.sh run suitable to your environment.  are these not sufficient?  if the underlying problem is an ability to set config options like these easily, we should address that.

Patch-set: 6