Update patch set 6
Patch Set 6: Code-Review+2

There is a real problem here that this fixes. I, for instance, couldn't start a cirros guest on a 3GB vagrant with 2 cpu on my laptop in a live demo because there wasn't 128 M of free memory left after all the python processes were running (and this was devstack default, so no swift, no ceilometer, no trove, no sahara).

I actually think a global dial down is really something we need like this, not just for the xenapi case.

Patch-set: 6
Label: Code-Review=+2