Update patch set 10
Patch Set 10:

Nadya, thank you for your comments.

Library for HBase written in an appropriate style to the a library for MySQL.
You can compare the function "recreate_nosql_database".
For SQL database types "lib/database" and for noSQL database types "lib/nosql-database".
Unlike SQL, i used "table_name" and "column_family".
Also variable "db_type" is used to indicate the type of the database and not as a name of data base.
This is due to the fact that the library "lib/nosql-database" not only for HBase but also for other noSQL databases.
like the "lib/database" not only for MySQL but also for PostgreSQL

As for the function "recreate_nosql_database" it is not used in DevStack and and it is never called.
But can be used for tests or in other libraries.
So after installation you get a clean HBASE without any data.

Thanks.

Patch-set: 10