Update patch set 1
Patch Set 1: Code-Review-1

I can see that a 1gb disk can't be snapshotted in 2gb drive, but does it need to go to 6?  wouldn't it be better to keep it close to the minimum (~3?) to flag unexpected growth.  why does it need to be the same as swift?

otoh, does this actually affect openstack infra usuage?  I'm not actually sure if there a ceph jobs?

Patch-set: 1
Label: Code-Review=-1