Replace screen_it() with run_process() throughout
run_process will use screen if USE_SCREEN=True (the default),
otherwise it will simply start the requested service. Therefore
wherever screen_it used, run_process can be instead.

To follow the precedent set in the parent
(0888a6213d9b701ac0aa037e35b637fa5943db93) places where screen_it
included an explicit `cd` to a base directory for the service have
been removed. While this may cut down on a measure of usefulness for
people doing development in the screens it protects against the use
of relative paths. This change may not be desired and can be removed
if necessary.

Where stop_screen was found it has been replaced with stop_process.

The template has been updated to reflect the use of the new
functions.

Change-Id: Ia3843818014f7c6c7526ef3aa9676bbddb8a85ca