Update patch set 9
Patch Set 9:

There's something not quite with run_process. When using it with nova-compute, exercises/volumes.sh fails in check-devstack-dsvm-cells because the scheduler is unaware of unavailable hosts to schedule. If the start for nova-compute is switched back to screen_it it works.

Committing that change (along with cleanups in variable use in run_process) now to see if check-devstack-dsvm-cells is able to finish clean or if there are more issues. Local testing suggests that cinder is going to be upset.

It's way past my bedtime so I need to stop, suggestions welcome.

Patch-set: 9
Reviewer: Gerrit User 11564 <11564@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0