Update patch set 9
Patch Set 9:

Zu Qiang:
Quite a long discussion i had on the IRC and understood the requirement for this automation. I still stand by the word that devstack installs perfectly with all proxy variables (http_proxy, https_proxy and no_proxy) exported manually. 

This patch (which is an extension of your patch mentioned above) was started with discussion on IRC with the community and i was informed i can extend it and proceed as there was no further update on your patch (Depending on the need).

If you wish to continue it, you can very well proceed and i can abandon this.

Reply for other comments are inline.
Thanks

Patch-set: 9