Update patch set 9
Patch Set 9:

hello, thanks for your response. i would like to continue with the work. 
however the problem i have is that the fix that you provided is only one part of my fix. in my fix, i also provide the fix when the devstack installation is called by other installation script, e.g. Trove. how to transfer the no_proxy setting out is a problem which you didn't provide any improvement. 
this is the reason that in my fix i have update the localrc file which people disagree. but without changing the lcoalrc, the no_proxy parameter cannot be transferred. the fix only can make the devstack installation works. but nothing more. 
if you like maybe we can work together on this issue.

Patch-set: 9