Update patch set 9
Patch Set 9:

Zu Qiang:
Happy to work together.

With respect to trove, do you feel the requirement of no_proxy to be carried to it. You can throw some light where would no_proxy needs to be used in trove.

I just tried to install trove-integration with the patch on devstack and i have not faced any error.

Patch-set: 9