Debian: use libvirt-bin when needed
libvirtd is the new name of the init script in Debian testing.
libvirt-bin is still in use on Debian Wheezy.

Since I222b71962f49896063910ff2a25e4f57be4bf819, libvirtd is the
default for Debian, this break the compatibility with Debian Wheezy.

With this patch, we use libvirt-bin only if there is no
/etc/init.d/libvirtd init script.

Change-Id: I13694fef93d36c2e128e15e7dbfaec9230335585