Ironic and baremetal install diskimage-builder with pip
DIB has recently started splitting out functionality into
other libraries and using requirements.txt to manage dependencies.
Devstack's master branch has dealt with this with the introduction
of a diskimage-builder service to handle properly installing from
git (topic:dib-gate).  This is the minimal required backport of that
and installs diskimage-builder from pip for ironic and baremetal.

Closes-bug: #1370759
Change-Id: Ic6613a11e2c4fb7a1261b81cdc2cf958515e60cd