Update patch set 1
Patch Set 1:

Is python-mysql.connector available in any of the Suse releases we attempted to support?  If so it would be more appropriate to use s #dist tag here to limit where it is attempted to be used.

Otherwise it would be polite to exit with an error when a configuration is attempted that we know works in other distros.  That isn't done more often for SuSE simply because most of us don't use it and don't know when that is the case.

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0