Update patch set 2
Patch Set 2: Code-Review-1

I noticed this week that we were logging the keystone tokens in the nova-api logs and the service_catalog gets logged in that, so yeah, it's huge already, we shouldn't add more stuff to it.

If the version discovery support is available and the keystone team can help getting that into novaclient it'd be helpful - I feel like there is a lot of 'there are better ways to do this' from the keystone team but there isn't much education on that - if it's all doc'ed somewhere can we be pointed to that?

Patch-set: 2