Update patch set 2
Patch Set 2:

@Brant

I guess your suggestion means

 Register endpoints without version part like:
  - http://%SERVICE_HOST%:8774/v2.0/$(tenant_id)s
  - http://%SERVICE_HOST%:8774/v2.1/$(tenant_id)s
  - http://%SERVICE_HOST%:8774/v3
  + http://%SERVICE_HOST%:8774/v3
 Client should generate endpoints from a version and endpoints which is getting from Keystone

right?

if so, it is difficult to do it now for Nova, because the return of "get version list" API doesn't include URL part of each version and we can customize URL point of versions by /etc/nova/api-paste.ini.

The return of "get version list" API
 $ curl -i 'http://192.168.11.62:8774/' -X GET -H "Accept: application/json" -H "X-Auth-Project-Id: demo" -H "X-Auth-Token: {SHA1}a478a30ec8bdadbdb5b8f98d97bf99ac83a8a1ea"
 [..]
 {"versions": [{"status": "CURRENT", "updated": "2011-01-21T11:33:21Z", "id": "v2.0", "links": [{"href": "http://192.168.11.62:8774/v2/", "rel": "self"}]}, {"status": "EXPERIMENTAL", "updated": "2013-07-23T11:33:21Z", "id": "v2.1", "links": [{"href": "http://192.168.11.62:8774/v2/", "rel": "self"}]}]}

The above "id" in the return shows version IDs and we can use them as URL part of versions if Nova environment is default setting. But we can change URL part of versions on /etc/nova/api-paste.ini and if doing that, we cannot use them as URL parts.

If the problem is increasing endpoints, how about removing current Nova v3 API endpoint?
Nova v3 API has disappeared on Juno cycle and we cannot use it now, so I think we can remove it.

Patch-set: 2