Update patch set 3
Patch Set 3: Code-Review+1

I agreed that swapping V3 for v2.1 is the pragmatic way to go for now.   

Going to full version discovery in the client worries me in terms of backwards compatibility -  if we go down that route we're going to have to do some careful thinking to make sure we don't break the client for existing deployed systems.

Patch-set: 3
Label: Code-Review=+1