Update patch set 3
Patch Set 3: Code-Review-1

The ideal is we register the compute endpoint as http://192.168.11.62:8774 (or, even more ideally, https://192.168.11.62/compute), then the client can query the endpoint, parse the response, figure out whether v2 or v2.1 are supported, and it also knows how to get to them because of the links.

This isn't something that Keystone is going to provide, it's going to have to be supported by the Nova server. We've got code in keystoneclient for doing discovery for the Identity API[1] that could be adapted to the compute API. Jamie Lennox has been working on this, and he might have plans to put this in the nova API, too. He's been working lately on providing a way to translate, for example, a catalog with a v2 compute endpoint to a v2.1 compute endpoint, see add_catalog_discover_hack() in [1].

I think removing the compute v3 endpoint should be in a separate commit. That's something we can all get behind.

[1] http://git.openstack.org/cgit/openstack/python-keystoneclient/tree/keystoneclient/discover.py

Patch-set: 3
Label: Code-Review=-1