Update patch set 3
Patch Set 3:

@Brant

I got your version-discovery idea, I think that is a right direction in long term.
and the *links* of "list versions" API response should work for it.

However, current *links* does not work for it because the one of v2.1 is the same as the one of v2.

 $ curl -i 'http://192.168.11.62:8774/' -X GET -H "Accept: application/json" -H "X-Auth-Project-Id: demo" -H "X-Auth-Token: {SHA1}a478a30ec8bdadbdb5b8f98d97bf99ac83a8a1ea"
 [..]
 {"versions": [
     {
        "status": "CURRENT", "updated": "2011-01-21T11:33:21Z", "id": "v2.0",
        "links": [{"href": "http://192.168.11.62:8774/v2/", "rel": "self"}]
     },
     {
        "status": "EXPERIMENTAL", "updated": "2013-07-23T11:33:21Z", "id": "v2.1",
        "links": [{"href": "http://192.168.11.62:8774/v2/", "rel": "self"}]
     }
 ]}

I'm not sure this is a bug of the API, maybe a bug..
In addition, we need to define the other endpoint name except "compute" because of backwards incompatibility as Phil said.
That means we need to consider novaclient behavior for using v2.1 API if version-discovery is necessary.
I feel now these works is a little overkill just for using v2.1 API.

The separated patches are reasonable, I will do it soon.

Patch-set: 3