Set mysql max_connections=200
There is a discrepancy between postgresql and mysql
max_connections configurations. postgresql max_connections
was set to 200 recently, while mysql remains at the default
of 151.

We remove the discrepancy with this commit. We also parametrize
the configuration in the same way for both mysql and postgres
through MAX_DB_CONNECTIONS.

Change-Id: I0682e222a3c5b4c8297372252f32ba291ae59fca