Expose MySQL max_connections setting
This patch exposes max_connections mysqld setting through
DATABASE_MAX_CONNECTIONS.

PostgreSQL backend does set max_connections to 200, and exposes it
with MAX_DB_CONNECTIONS, while default for MySQL is 151. So we can not
just set both to the same value without changing one of them.
The same name is not used to avoid initial conflict with PostgreSQL and
to maintain naming scheme that already exists in stackrc.

PostgreSQL backend would be updated to respect the value of
DATABASE_MAX_CONNECTIONS in a followup.

Change-Id: I0682e222a3c5b4c8297372252f32ba291ae59fca