Update patch set 1
Patch Set 1:

The token for the same user it got changed in between.
As you can see the bug. It is failing in fedora image only and on debugging in glance service logs, got to know that token do not exists. So reading the new token to make it run.
Change has been tested.

Patch-set: 1