Update patch set 1
Patch Set 1:

Dean,
The issues is that this discrepancy might be causing some of the tests to intermittently fail for mysql, while they do not fail for postgres. I did not see harm in making both mysql and postgres max_connections to be the same by raising the mysql limit. Please correct me if I am wrong.

Also, exposing the configuration of mysql max_connections would allow third party CI to configure it (addition required in devstack-gate for it to work).

Patch-set: 1
Label: Verified=0