Use dib service to build Ironic pxe_ssh deploy ramdisk
Rely on the the DIB service's functionality for building the Ironic
pxe_ssh ramdisk. This removes fragmentation and allows the DIB environment
to be configured and managed in a single place.

In addition to calling ramdisk_image_create, this moves validation of required
services to earlier in execution and adds a check for the dib service being
enabled.

Change-Id: Idfbcc89375f5c84fae18b31027cda44ba76e0e2d
Partial-bug: #1375488