Add RAW image support for Ceph
To work in ideal circonstances, Ceph needs a RAW image in Glance. This
allows either Nova or Cinder to take advantage of the copy-on-write
cloning functionnality. This commit introduces a generic change_img_format
function that can be used by any component.
The default behavior will now download the cirros image and will convert
it into RAW.

Change-Id: I0d1157829f7131925d2d6918e8bb0f7cebe1e726
Signed-off-by: Sébastien Han <sebastien.han@enovance.com>