Update patch set 2
Patch Set 2:

So I may have misunderstood exactly what this does.

The indention on the new block of code is off.  That is what made me think it was inserted in the wrong place.  

I don't understand why DATABASE_HOST is tested and debconf set up only if .my.cnf doesn't exist...

Patch-set: 2