Update patch set 1
Patch Set 1:

I'm either running it wrong or bash8 is not properly checking what it says it will. The docs say that a function of the form "foo()" instead of "function foo {" ought to be an error, but it's not reporting that.

However I didn't make the change to make bash8 happy but rather to make the files internally consistent with themselves. Everywhere else the ^function* form is used.

Patch-set: 1
Reviewer: Gerrit User 11564 <11564@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0