Update patch set 1
Patch Set 1:

Thanks Eoghan. I considered automatically starting multiples of the various coordinating agents but then decided not to for a few different reasons:

* managing the process and/or screens gets messy (see how lib/nova handles multiple fake rivers for an example)
* in the multinode case, we don't want to arbitrarily choose to have multiple agents on _this_ node when perhaps we want one on _this_ node and one on each of those _other_ nodes
* someone might want multiples of one type of agent and not the other

None of that stuff can be controlled just by the existence of the one variable, we'd need more.

So in order to maintain some simplicity I thought best to start the idea with a small change and then see where it takes us.

None of the above issues are insurmountable (have a problem in devstack just add more VARIABLES!) so if you think adding more features in this patch is the way to go, shout out.

For reference, the specific use case of mine that I'm trying to cover here is wanting a devstack that is already ready for coordination and then I, by hand, start and stop other agents to add to and remove from the various groups. That is I'm thinking of this as a development and testing aid, one that involves humans exploring, not automation.

Patch-set: 1
Label: Verified=0